Moved CrossrefAdapter and CrossrefExtension into lib/rdoc/generator/markdown/crossref.rb.
